+++ src/PublicRestApi/Rooms/RoomsApi.php
@@ -5,23 +5,24 @@ declare(strict_types=1);
 namespace Dbp\CampusonlineApi\PublicRestApi\Rooms;
 
 use Dbp\CampusonlineApi\Helpers\ApiException;
-use Dbp\CampusonlineApi\PublicRestApi\Api;
+use Dbp\CampusonlineApi\PublicRestApi\AbstractApi;
 use Dbp\CampusonlineApi\Rest\Tools;
 use GuzzleHttp\Exception\GuzzleException;
 use Psr\Http\Message\ResponseInterface;
 
-class RoomsApi extends Api
+class RoomsApi extends AbstractApi
 {
-    private const API_PATH = 'co/co-brm-core/facilities/api/rooms';
+    private const API_PATH = 'co/co-brm-core/facilities/api';
+    private const ROOMS_API_PATH = self::API_PATH.'/rooms';
 
     private const ROOM_UID_QUERY_PARAMETER_NAME = 'room_uid';
 
-    public function getRoomByIdentifier(string $identifier): RoomResource
+    public function getRoomByIdentifier(int $identifier): RoomResource
     {
         try {
             $roomResources = iterator_to_array($this->getRoomsFromResponse(
                 $this->connection->getClient()->get(
-                    self::API_PATH.'?'.http_build_query([
+                    self::ROOMS_API_PATH.'?'.http_build_query([
                         self::ROOM_UID_QUERY_PARAMETER_NAME => $identifier,
                     ])
                 )));
@@ -38,12 +39,17 @@ class RoomsApi extends Api
     /**
      * @return iterable<RoomResource>
      */
-    public function getRooms(int $firstItemIndex, int $maxNumItems, array $options = []): iterable
+    public function getRooms(int $firstItemIndex = 0, int $maxNumItems = 30, array $options = []): iterable
     {
         try {
-            return $this->getRoomsFromResponse(
-                $this->connection->getClient()->get(self::API_PATH.'?'.
-                    http_build_query(self::getPaginationQueryParameters($firstItemIndex, $maxNumItems))));
+            // WORKAROUND: CO ignores limit=0
+            if ($maxNumItems === 0) {
+                return [];
+            } else {
+                return $this->getRoomsFromResponse(
+                    $this->connection->getClient()->get(self::ROOMS_API_PATH.'?'.
+                        http_build_query(self::getOffsetBasedPaginationQueryParameters($firstItemIndex, $maxNumItems))));
+            }
         } catch (GuzzleException $guzzleException) {
             throw ApiException::fromGuzzleException($guzzleException);
         }